+++ src/tileFormats/PntsValidator.ts
@@ -92,6 +92,8 @@ const featureTableSemantics = {
 /**
  * A class that can perform validation of PNTS data that is
  * given as a Buffer.
+ *
+ * @internal
  */
 export class PntsValidator implements Validator<Buffer> {
   async validateObject(