+++ src/tileFormats/legacy/validateFeatureTable.ts
@@ -14,6 +14,8 @@ import { componentTypeToByteLength } from "./utility";
  * @param featuresLength - The number of features.
  * @param featureTableSemantics - An object containing semantic information for each feature table property, specific to the tile format.
  * @returns An error message if validation fails, otherwise undefined.
+ *
+ * @internal
  */
 function validateFeatureTable(
   featureTableJson: any,