+++ src/traversal/TilesetTraverser.ts
@@ -11,6 +11,8 @@ import { Schema } from "../structure/Metadata/Schema";
 
 /**
  * A class that can traverse the tiles of a tileset.
+ *
+ * @internal
  */
 export class TilesetTraverser {
   /**