+++ src/validation/ExtendedObjectsValidators.ts
@@ -16,15 +16,25 @@ import { SemanticValidationIssues } from "../issues/SemanticValidationIssues";
  * extensions). When an object contains an extension with one
  * of the registered names, then the respective validators will
  * be applied to that object.
+ *
+ * @internal
  */
 export class ExtendedObjectsValidators {
   /**
    * The mapping from extension names to the validators that
    * are used for objects that contain the respective extension.
    */
-  static readonly extendedObjectValidators = new Map<string, Validator<any>>();
+  private static readonly extendedObjectValidators = new Map<
+    string,
+    Validator<any>
+  >();
 
-  static readonly overrides = new Map<string, boolean>();
+  /**
+   * The mapping from extension names to the flag that indicates
+   * whether the corresponding validator should override the
+   * default validation process.
+   */
+  private static readonly overrides = new Map<string, boolean>();
 
   /**
    * Registers a validator for an object with the specified extension.