+++ src/validation/ValidationContext.ts
@@ -1,3 +1,5 @@
+import { defaultValue } from "../base/defaultValue";
+
 import { ResourceResolver } from "../io/ResourceResolver";
 
 import { ValidationIssue } from "./ValidationIssue";
@@ -21,6 +23,7 @@ import { ValidationResult } from "./ValidationResult";
  * of the validator (for example, whether external resources should
  * be validated) are stored in a `ValidationOptions` object.
  *
+ * @internal
  */
 export class ValidationContext {
   /**
@@ -45,8 +48,8 @@ export class ValidationContext {
    */
   private readonly _resourceResolver: ResourceResolver;
 
-  constructor(resourceResolver: ResourceResolver) {
-    this._options = new ValidationOptions();
+  constructor(resourceResolver: ResourceResolver, options?: ValidationOptions) {
+    this._options = defaultValue(options, new ValidationOptions());
     this._result = ValidationResult.create();
     this._resourceResolver = resourceResolver;
     this._extensionsFound = new Set<string>();
@@ -80,8 +83,7 @@ export class ValidationContext {
   deriveFromResourceResolver(
     resourceResolver: ResourceResolver
   ): ValidationContext {
-    const derived = new ValidationContext(resourceResolver);
-    derived._options = this._options;
+    const derived = new ValidationContext(resourceResolver, this._options);
     derived._extensionsFound = this._extensionsFound;
     return derived;
   }